resource-timing/iframe-refresh.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing iframe refresh</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    let t = async_test("Test that iframe refreshes are not observable by the parent");
    const pre_refresh_url = new URL("resources/document-that-refreshes.html", location).href;
    const post_refresh_url = new URL("resources/document-refreshed.html", location).href;
    function verify_iframe_refresh_not_observable() {
        let entries = performance.getEntriesByType("resource");
        let found_first_iframe = false;
        for (entry of entries) {
            if (entry.name == pre_refresh_url) {
                found_first_iframe = true;
            }
            assert_not_equals(entry.name, post_refresh_url, "Refreshed iframe should not be observable");
        }
        assert_true(found_first_iframe, "Initial iframe should be observable");
        t.done();
    }
    window.addEventListener("message", t.step_func(e=>{
        if (e.data == "refreshed") {
            verify_iframe_refresh_not_observable();
        }
    }));
</script>
<iframe></iframe>
<script>
    document.getElementsByTagName("iframe")[0].src = pre_refresh_url;
</script>
